refactor: transform internet module
