+++ src/modules/internet/index.ts
@@ -1,1 +1,6 @@
+export type { EmojiType } from './emoji';
+export type { HTTPStatusCodeType } from './http-status-code';
+export { IPv4Network } from './ipv4';
+export type { IPv4NetworkType } from './ipv4';
 export * from './module';
+export type { HTTPProtocolType } from './url';
