+++ src/modules/internet/mac.ts
@@ -67,7 +67,6 @@ export function mac(
         separator?: string;
       }
 ): string;
-
 export function mac(
   fakerCore: FakerCore,
   options:
